% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Datasets}
\alias{Datasets}
\alias{humidity_max}
\alias{humidity_min}
\title{Maximum and minimum values of daily relative air humidity in the Atacama Desert}
\usage{
data(humidity_max)
data(humidity_min)
}
\description{
Maximum and minimum values of daily relative air humidity in the Atacama Desert
}
\author{
José G S Sena, Fabio M Bayer, Paulo H. Ferreira, José A Fiorucci
}
\keyword{data}
